moved RinchiFullInfo.java to private static class in JnaRinchiTest
